+++ react/HasDifferentPrices.tsx
@@ -1,0 +1,3 @@
+import HasDifferentPrices from './components/HasDifferentPrices'
+
+export default HasDifferentPrices